move translation presentation into separate file
